anvil-weather-proxy/requirements.txt
# FastPlanner Weather Proxy - Anvil Requirements
# These packages are needed for the weather proxy server

# Core dependencies
requests>=2.28.0
anvil-uplink>=0.4.0

# Optional for enhanced functionality
python-dateutil>=2.8.0
pytz>=2022.1